Use thiserror to create a specific error type for the coordinator
Rather than bailing with a string error (e.g. in the "no complete timestamps" case) we can create a custom error enum using `thiserror` and a variant that covers that specific error case.

We can then later match against that specific variant, and can also carry along any relevant metadata that might help the user with resolving the error (or might indicator to the system that the error is transient and perhaps should not even lead to the termination of the program).